arroy

Module distances

The set of distances implementing the Distance and supported by arroy.

Enums

BinaryQuantizedCosine

The Cosine similarity is a measure of similarity between two non-zero vectors defined in an inner product space. Cosine similarity is the cosine of the angle between the vectors. /!\ This distance function is binary quantized, which means it loses all its precision and their scalar values are converted to -1 or 1.

BinaryQuantizedEuclidean

The Euclidean distance between two points in Euclidean space is the length of the line segment between them.

BinaryQuantizedManhattan

A taxicab geometry or a Manhattan geometry is a geometry whose usual distance function or metric of Euclidean geometry is replaced by a new metric in which the distance between two points is the sum of the absolute differences of their Cartesian coordinates. /!\ This distance function is binary quantized, which means it loses all its precision and their scalar values are converted to -1 or 1.

Cosine

The Cosine similarity is a measure of similarity between two non-zero vectors defined in an inner product space. Cosine similarity is the cosine of the angle between the vectors.

DotProduct

In mathematics, the dot product or scalar product is an algebraic operation that takes two equal-length sequences of numbers (usually coordinate vectors), and returns a single number.

Euclidean

The Euclidean distance between two points in Euclidean space is the length of the line segment between them.

Manhattan

A taxicab geometry or a Manhattan geometry is a geometry whose usual distance function or metric of Euclidean geometry is replaced by a new metric in which the distance between two points is the sum of the absolute differences of their Cartesian coordinates.