arroy

Module internals

The set of types used by the Distance trait.

Structs

Leaf

A leaf node which corresponds to the vector inputed by the user and the distance header.

NodeCodec

The codec used internally to encode and decode nodes.

NodeHeaderBinaryQuantizedCosine

The header of BinaryQuantizedCosine leaf nodes.

NodeHeaderBinaryQuantizedEuclidean

The header of BinaryQuantizedEuclidean leaf nodes.

NodeHeaderBinaryQuantizedManhattan

The header of BinaryQuantizedEuclidean leaf nodes.

NodeHeaderCosine

The header of Cosine leaf nodes.

NodeHeaderDotProduct

The header of DotProduct leaf nodes.

NodeHeaderEuclidean

The header of Euclidean leaf nodes.

NodeHeaderManhattan

The header of Manhattan leaf nodes.

SizeMismatch

Returned in case you tried to make an unaligned vector from a slice of bytes that don’t have the right number of elements

UnalignedVector

A wrapper struct that is used to read unaligned vectors directly from memory.

Enums

KeyCodec

The heed codec used internally to encode/decoding the internal key type.

Side

A type that is used to decide on which side of a plane we move an item.

Traits

UnalignedVectorCodec

Determine the way the vectors should be read and written from the database