arroy::internals

Struct UnalignedVector

pub struct UnalignedVector<Codec: UnalignedVectorCodec> { /* private fields */ }

A wrapper struct that is used to read unaligned vectors directly from memory.

Implementations

impl<Codec: UnalignedVectorCodec> UnalignedVector<Codec>

pub fn from_bytes( bytes: &[u8], ) -> Result<Cow<'_, UnalignedVector<Codec>>, SizeMismatch>

Creates an unaligned vector from a slice of bytes. Don’t allocate.

pub fn from_slice(slice: &[f32]) -> Cow<'_, UnalignedVector<Codec>>

Creates an unaligned vector from a slice of f32. May allocate depending on the codec.

pub fn from_vec(vec: Vec<f32>) -> Cow<'static, UnalignedVector<Codec>>

Creates an unaligned slice of f32 wrapper from a slice of f32. The slice is already known to be of the right length.

pub fn iter(&self) -> impl ExactSizeIterator<Item = f32> + '_

Returns an iterator of f32 that are read from the vector. The f32 are copied in memory and are therefore, aligned.

pub fn is_zero(&self) -> bool

Returns true if all the elements in the vector are equal to 0.

pub fn to_vec(&self) -> Vec<f32>

Returns an allocated and aligned Vec<f32>.

pub fn len(&self) -> usize

Returns the len of the vector in terms of elements.

pub fn is_empty(&self) -> bool

Returns wether it is empty or not.