arroy::internals

Trait UnalignedVectorCodec

pub trait UnalignedVectorCodec: ToOwned + Sized {
    // Required methods
    fn from_bytes(
        bytes: &[u8],
    ) -> Result<Cow<'_, UnalignedVector<Self>>, SizeMismatch>;
    fn from_slice(slice: &[f32]) -> Cow<'_, UnalignedVector<Self>>;
    fn from_vec(vec: Vec<f32>) -> Cow<'static, UnalignedVector<Self>>;
    fn to_vec(vec: &UnalignedVector<Self>) -> Vec<f32>;
    fn iter(
        vec: &UnalignedVector<Self>,
    ) -> impl ExactSizeIterator<Item = f32> + '_;
    fn len(vec: &UnalignedVector<Self>) -> usize;
    fn is_zero(vec: &UnalignedVector<Self>) -> bool;
}

Determine the way the vectors should be read and written from the database

Required Methods

fn from_bytes( bytes: &[u8], ) -> Result<Cow<'_, UnalignedVector<Self>>, SizeMismatch>

Creates an unaligned vector from a slice of bytes. Don’t allocate.

fn from_slice(slice: &[f32]) -> Cow<'_, UnalignedVector<Self>>

Creates an unaligned vector from a slice of f32. May allocate depending on the codec.

fn from_vec(vec: Vec<f32>) -> Cow<'static, UnalignedVector<Self>>

Creates an unaligned slice of f32 wrapper from a slice of f32. The slice is already known to be of the right length.

fn to_vec(vec: &UnalignedVector<Self>) -> Vec<f32>

Converts the UnalignedVector to an aligned vector of f32. It’s strictly equivalent to .iter().collect() but the performances are better.

fn iter(vec: &UnalignedVector<Self>) -> impl ExactSizeIterator<Item = f32> + '_

Returns an iterator of f32 that are read from the vector. The f32 are copied in memory and are therefore, aligned.

fn len(vec: &UnalignedVector<Self>) -> usize

Returns the len of the vector in terms of elements.

fn is_zero(vec: &UnalignedVector<Self>) -> bool

Returns true if all the elements in the vector are equal to 0.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.