arroy

Struct QueryBuilder

pub struct QueryBuilder<'a, D: Distance> { /* private fields */ }

Options used to make a query against an arroy Reader.

Implementations

impl<'a, D: Distance> QueryBuilder<'a, D>

pub fn by_item( &self, rtxn: &RoTxn<'_>, item: ItemId, ) -> Result<Option<Vec<(ItemId, f32)>>>

Returns the closests items from item.

See also Self::by_vector.

Examples

reader.nns(20).by_item(&rtxn, 5);

pub fn by_vector( &self, rtxn: &RoTxn<'_>, vector: &'a [f32], ) -> Result<Vec<(ItemId, f32)>>

Returns the closest items from the provided vector.

See also Self::by_item.

Examples

reader.nns(20).by_vector(&rtxn, &[1.25854, -0.75598, 0.58524]);

pub fn search_k(&mut self, search_k: NonZeroUsize) -> &mut Self

During the query, arroy will inspect up to search_k nodes which defaults to n_trees * count if not provided. search_k gives you a run-time tradeoff between better accuracy and speed.

Examples

use std::num::NonZeroUsize;
reader.nns(20).search_k(NonZeroUsize::new(1000).unwrap()).by_item(&rtxn, 3);

pub fn oversampling(&mut self, oversampling: NonZeroUsize) -> &mut Self

Oversampling will multiply [search_k] by the specified number. That’s useful when you don’t want to compute search_k yourself.

Examples

use std::num::NonZeroUsize;
reader.nns(20).oversampling(NonZeroUsize::new(6).unwrap()).by_item(&rtxn, 5);

pub fn candidates(&mut self, candidates: &'a RoaringBitmap) -> &mut Self

Specify a subset of candidates to inspect. Filters out everything else.

Examples

let candidates = roaring::RoaringBitmap::from_iter([1, 3, 4, 5, 6, 7, 8, 9, 15, 16]);
reader.nns(20).candidates(&candidates).by_item(&rtxn, 6);